replace outdated location module
